Refactored event.
